Added logging `log()` function override to library
By overriding the std.log `log` function in the root of your program,
you can intercept log writes. This commit allows users to override that
function in their root source file, but pass the parameters to the
function available in the library's root.

This makes logging to files easy for users of the library, and
furthermore, the library can use logging wherever needed, and it only
passes through whatever the user has overridden the `log()` function.